requirements.txt
# XRPL Dependencies
xrpl-py==3.0.0

# Trading & Market Data
ccxt==4.4.36
pandas==2.2.2
numpy==1.26.4
ta==0.11.0

# Async & Concurrency
asyncio-mqtt==0.16.2
aiohttp==3.10.11

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Logging & Monitoring
loguru==0.7.3
prometheus-client==0.21.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
requests==2.32.3

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2

# Visualization
matplotlib==3.9.3